@@ -9,15 +9,15 @@
     <packaging>war</packaging>
 
     <properties>
-        <java.version>11</java.version>
-        <hilla.version>1.3.0</hilla.version>
-        <selenium.version>4.4.0</selenium.version>
+        <java.version>17</java.version>
+        <hilla.version>2.3.4</hilla.version>
+        <!-- <selenium.version>4.14.0</selenium.version> -->
     </properties>
 
     <parent>
         <groupId>org.springframework.boot</groupId>
         <artifactId>spring-boot-starter-parent</artifactId>
-        <version>2.7.3</version>
+        <version>3.1.1</version>
     </parent>
 
     <repositories>
@@ -122,7 +122,7 @@
         <dependency>
             <groupId>com.vaadin</groupId>
             <artifactId>exampledata</artifactId>
-            <version>6.0.0</version>
+            <version>6.2.0</version>
         </dependency>
 
         <dependency>
@@ -138,23 +138,23 @@
         <dependency>
             <groupId>org.springframework.boot</groupId>
             <artifactId>spring-boot-starter-validation</artifactId>
-            <exclusions>
+            <!-- <exclusions>
                 <exclusion>
                     <groupId>org.apache.tomcat.embed</groupId>
                     <artifactId>tomcat-embed-el</artifactId>
                 </exclusion>
-            </exclusions>
+            </exclusions> -->
         </dependency>
         <dependency>
-            <groupId>javax.el</groupId>
-            <artifactId>javax.el-api</artifactId>
-            <version>3.0.0</version>
+            <groupId>jakarta.el</groupId>
+            <artifactId>jakarta.el-api</artifactId>
+            <version>5.0.0</version>
         </dependency>
-        <dependency>
+        <!-- <dependency>
             <groupId>org.glassfish</groupId>
             <artifactId>javax.el</artifactId>
             <version>3.0.0</version>
-        </dependency>
+        </dependency> -->
         <dependency>
             <groupId>org.springframework.boot</groupId>
             <artifactId>spring-boot-devtools</artifactId>
@@ -168,6 +168,7 @@
         <dependency>
             <groupId>com.vaadin</groupId>
             <artifactId>vaadin-testbench</artifactId>
+            <version>24.2.5</version>
             <scope>test</scope>
         </dependency>
         <!-- Include JUnit 4 support for TestBench and others -->
@@ -185,7 +186,7 @@
         <dependency>
             <groupId>io.github.bonigarcia</groupId>
             <artifactId>webdrivermanager</artifactId>
-            <version>5.1.1</version>
+            <version>5.4.1</version>
             <scope>test</scope>
         </dependency>
     </dependencies>  
