@@ -58,11 +58,11 @@ export class Lang {
             "offline": "Offline",
             "online": "Online",
             "datepicker-format": "date format is not correct",
-            "validationError.PastOrPresentWeekdayAndRequired": "date is required, date can't be in future, date must be weekday",
-            "validationError.NotEmpty": "input is required",
+            "validationError.date input is not valid": "date is required, date can't be in future, date must be weekday",
+            "validationError.must not be empty": "input is required",
             "validationError.NotNull": "input is required",
-            "validationError.Email": "input is not valid e-mail address",
-            "validationError.PositiveOrZero": "value can't be negative",
+            "validationError.must be a well-formed email address": "input is not valid e-mail address",
+            "validationError.must be greater than or equal to 0": "value can't be negative",
             "login-title": "Log in",
             "login-username": "Username",
             "login-password": "Password",
@@ -118,11 +118,11 @@ export class Lang {
             "offline": "Ei verkkoyhteyttä",
             "online": "Verkossa",
             "datepicker-format": "päivämäärän muotoilu ei ole oiken",
-            "validationError.PastOrPresentWeekdayAndRequired": "päivämäärä vaadittu, päivämämäärä ei saa tulevaisuudessa, päivämäärä ei saa olla viikonloppuna",
-            "validationError.NotEmpty": "syöte ei saa olla tyhjä",
+            "validationError.date input is not valid": "päivämäärä vaadittu, päivämämäärä ei saa tulevaisuudessa, päivämäärä ei saa olla viikonloppuna",
+            "validationError.must not be empty": "syöte ei saa olla tyhjä",
             "validationError.NotNull": "syöte ei saa olla tyhjä",
-            "validationError.Email": "syöte ei ole kelvollinen sähköpostiosoite",
-            "validationError.PositiveOrZero": "arvo ei saa olla negatiivinen",
+            "validationError.must be a well-formed email address": "syöte ei ole kelvollinen sähköpostiosoite",
+            "validationError.must be greater than or equal to 0": "arvo ei saa olla negatiivinen",
             "login-title": "Kirjautuminen",
             "login-username": "Käyttäjä",
             "login-password": "Salasana",
